board_data: separate the build generator from the board structure
